feat: separate render dialog to AppRenderDialog
